refactor: split pandas system tests to new module
